Refactor into a single module
