+++ src/fontra/views/editor/editor.html
@@ -12,7 +12,6 @@
     <script type="module" src="/web-components/grouped-settings.js"></script>
     <script type="module" src="/web-components/inline-svg.js"></script>
     <script type="module" src="/web-components/modal-dialog.js"></script>
-    <script type="module" src="/web-components/reference-font.js"></script>
     <script type="module" src="/web-components/ui-list.js"></script>
   </head>
   <body>
