+++ src/fontra/views/editor/panel.js
@@ -1,9 +1,15 @@
 import { SimpleElement } from "/core/unlit.js";
 
 export default class Panel extends SimpleElement {
-  constructor(editorController) {
+  constructor(editorController, renderContentElement = true) {
     super();
     this.editorController = editorController;
+    if (renderContentElement) {
+      this.renderContentElement();
+    }
+  }
+
+  renderContentElement() {
     this.contentElement = this.getContentElement();
     this.shadowRoot.appendChild(this.contentElement);
   }
